Move script utility functions to utils.ts and provide tests for them
