+++ scripts/csv-to-couchdb.ts
@@ -3,29 +3,9 @@
 import Denomander from "https://deno.land/x/denomander@0.9.1/mod.ts";
 import { readCSV } from "https://deno.land/x/csv/mod.ts";
 import ProgressBar from "https://deno.land/x/progress@v1.2.4/mod.ts";
-import { BufReader } from "https://deno.land/std/io/buffer.ts";
 import { createDatabase, postBulkDocs } from "../pkg/couchdb/couchdb.ts";
+import { inChunks, rowToObjectMaker, howManyLines } from "./utils.ts";
 
-/**
- * Counts lines in a csv file for progress bar calculations
- * @param {string} filePath - csv file path
- * @return {Number} - Number of lines in file
- */
-
-const howManyLines = async function (filePath: string) {
-  let lineCount: number = 0;
-  var file;
-  try {
-    file = await Deno.open(filePath);
-    const bufReader = new BufReader(file);
-    while (await bufReader.readLine()) {
-      lineCount++;
-    }
-  } finally {
-    if (file) file.close();
-  }
-  return lineCount;
-};
 
 //#region progress bar
 /**
@@ -132,40 +112,3 @@ program.command("parse", "parses csv file into the db")
   .parse(Deno.args);
 
 //#endregion program
-
-//#region chunk generator
-
-/**
- * Aggregate values from an Iterable in chunks of a given maximum size
- * @param {Iterator} sequence The iterator to take values from
- * @param {Number} chunkSize - The size of the chunks
- * @return {Iterator} - An iterator over the chunks
- */
-const inChunks = async function* <T>(
-  sequence: AsyncIterable<AsyncIterable<string>>,
-  chunkSize: number,
-) {
-  let chunk = [];
-  for await (const element of sequence) {
-    chunk.push(element);
-    if (chunk.length === chunkSize) {
-      yield chunk;
-      chunk = [];
-    }
-    yield chunk;
-  }
-};
-
-
-const rowToObjectMaker = function(columns: string[]) {
-  return async function(row: any) {
-    let obj: { [key: string]: string } = {};
-    let i = 0;
-    for await (const field of row) {
-      obj[columns[i]] = field;
-      i++;
-    }
-    return obj;
-  };
-}
-//#endregion chunk generator
